+++ SonicStreamer/Pages/ArtistsPage.xaml.cs
@@ -1,31 +1,36 @@
-﻿using SonicStreamer.Common.System;
+﻿using System;
+using SonicStreamer.Common.System;
 using SonicStreamer.ViewModels;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Navigation;
+using SonicStreamer.Controls;
 
 namespace SonicStreamer.Pages
 {
     public sealed partial class ArtistsPage : Page
     {
-        readonly ArtistsViewModel _artistVm;
-        readonly PlaylistViewModel _playlistVm;
+        public readonly ArtistsViewModel ArtistVm;
+        public readonly PlaylistViewModel PlaylistVm;
+        public readonly MainViewModel MainVm;
 
         public ArtistsPage()
         {
             InitializeComponent();
 
-            if (ResourceLoader.Current.GetResource(ref _artistVm, Constants.ViewModelArtist) == false)
-                _artistVm = new ArtistsViewModel();
-            if (ResourceLoader.Current.GetResource(ref _playlistVm, Constants.ViewModelPlaylist) == false)
-                _playlistVm = new PlaylistViewModel();
+            if (ResourceLoader.Current.GetResource(ref ArtistVm, Constants.ViewModelArtist) == false)
+                ArtistVm = new ArtistsViewModel();
+            if (ResourceLoader.Current.GetResource(ref PlaylistVm, Constants.ViewModelPlaylist) == false)
+                PlaylistVm = new PlaylistViewModel();
+            if (ResourceLoader.Current.GetResource(ref MainVm, Constants.ViewModelMain) == false)
+                MainVm = new MainViewModel();
         }
 
         protected override async void OnNavigatedTo(NavigationEventArgs e)
         {
             Microsoft.HockeyApp.HockeyClient.Current.TrackPageView(GetType().Name);
-            await _artistVm.LoadDataAsync();
-            ListViewSource.Source = _artistVm.Items;
+            await ArtistVm.LoadDataAsync();
+            ListViewSource.Source = ArtistVm.Items;
             var listViewBase = SemanticZoomContainer.ZoomedOutView as ListViewBase;
             if (listViewBase != null)
                 listViewBase.ItemsSource = ListViewSource.View.CollectionGroups;
@@ -36,19 +41,21 @@ namespace SonicStreamer.Pages
             Frame.Navigate(typeof(TrackListingPage), e.ClickedItem);
         }
 
-        private async void PlaylistFlyout_Opening(object sender, object e)
+        private async void AddToPlaylist_Click(object sender, RoutedEventArgs e)
         {
-            await _playlistVm.LoadFlyoutDataAsync();
+            AddToPlaylistDialog.Content = new AddToPlaylistDialog();
+
+            var dialogResult = await AddToPlaylistDialog.ShowAsync();
+            if (dialogResult == ContentDialogResult.Primary)
+            {
+                await ArtistVm.AddToPlaylistAsync();
+            }
+            PlaylistVm.ResetDialogInputs();
         }
 
-        private void PlaylistFlyout_Closed(object sender, object e)
+        private async void AddToPlaylistDialog_OnOpened(ContentDialog sender, ContentDialogOpenedEventArgs args)
         {
-            _playlistVm.ResetFlyoutInputs();
-        }
-
-        private void AddToPlayback_Click(object sender, RoutedEventArgs e)
-        {
-            PlaylistFlyout.Hide();
+            await PlaylistVm.LoadDialogDataAsync();
         }
     }
 }