+++ SonicStreamer/Pages/SettingsPage.xaml.cs
@@ -1,18 +1,30 @@
 ﻿using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Navigation;
+using SonicStreamer.Common.System;
+using SonicStreamer.ViewModels;
 
 namespace SonicStreamer.Pages
 {
     public sealed partial class SettingsPage : Page
     {
+        public readonly ApplicationSettingsViewModel SettingsVm;
+        public readonly MainViewModel MainVm;
+
         public SettingsPage()
         {
             InitializeComponent();
+
+            if (ResourceLoader.Current.GetResource(ref SettingsVm, Constants.ViewModelApplicationSettings) == false)
+                SettingsVm = new ApplicationSettingsViewModel();
+            if (ResourceLoader.Current.GetResource(ref MainVm, Constants.ViewModelMain) == false)
+                MainVm = new MainViewModel();
         }
 
-        protected override void OnNavigatedTo(NavigationEventArgs e)
+        protected override async void OnNavigatedTo(NavigationEventArgs e)
         {
             Microsoft.HockeyApp.HockeyClient.Current.TrackPageView(GetType().Name);
+
+            await SettingsVm.LoadDataAsync();
         }
     }
 }