+++ SonicStreamer/Pages/TrackListingPage.xaml.cs
@@ -1,28 +1,30 @@
-﻿using SonicStreamer.Common.System;
+﻿using System;
+using SonicStreamer.Common.System;
 using SonicStreamer.Subsonic.Data;
 using SonicStreamer.ViewModels;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Navigation;
+using SonicStreamer.Controls;
 
 namespace SonicStreamer.Pages
 {
     public sealed partial class TrackListingPage : Page
     {
-        readonly TrackListingViewModel _trackListingVm;
-        readonly MainViewModel _mainVm;
-        readonly PlaylistViewModel _playlistVm;
+        public readonly TrackListingViewModel TrackListingVm;
+        public readonly MainViewModel MainVm;
+        public readonly PlaylistViewModel PlaylistVm;
 
         public TrackListingPage()
         {
             InitializeComponent();
 
-            if (ResourceLoader.Current.GetResource(ref _trackListingVm, Constants.ViewModelTrackListing) == false)
-                _trackListingVm = new TrackListingViewModel();
-            if (ResourceLoader.Current.GetResource(ref _mainVm, Constants.ViewModelMain) == false)
-                _mainVm = new MainViewModel();
-            if (ResourceLoader.Current.GetResource(ref _playlistVm, Constants.ViewModelPlaylist) == false)
-                _playlistVm = new PlaylistViewModel();
+            if (ResourceLoader.Current.GetResource(ref TrackListingVm, Constants.ViewModelTrackListing) == false)
+                TrackListingVm = new TrackListingViewModel();
+            if (ResourceLoader.Current.GetResource(ref MainVm, Constants.ViewModelMain) == false)
+                MainVm = new MainViewModel();
+            if (ResourceLoader.Current.GetResource(ref PlaylistVm, Constants.ViewModelPlaylist) == false)
+                PlaylistVm = new PlaylistViewModel();
         }
 
         protected override async void OnNavigatedTo(NavigationEventArgs e)
@@ -31,23 +33,25 @@ namespace SonicStreamer.Pages
             var param = e.Parameter as SubsonicMusicObject;
             if (param != null)
             {
-                await _trackListingVm.LoadDataAsync(param);
+                await TrackListingVm.LoadDataAsync(param);
             }
         }
 
-        private async void PlaylistFlyout_Opening(object sender, object e)
+        private async void AddToPlaylist_Click(object sender, RoutedEventArgs e)
         {
-            await _playlistVm.LoadFlyoutDataAsync();
+            AddToPlaylistDialog.Content = new AddToPlaylistDialog();
+
+            var dialogResult = await AddToPlaylistDialog.ShowAsync();
+            if (dialogResult == ContentDialogResult.Primary)
+            {
+                await TrackListingVm.AddToPlaylistAsync();
+            }
+            PlaylistVm.ResetDialogInputs();
         }
 
-        private void PlaylistFlyout_Closed(object sender, object e)
+        private async void AddToPlaylistDialog_OnOpened(ContentDialog sender, ContentDialogOpenedEventArgs args)
         {
-            _playlistVm.ResetFlyoutInputs();
-        }
-
-        private void AddToPlayback_Click(object sender, RoutedEventArgs e)
-        {
-            PlaylistFlyout.Hide();
+            await PlaylistVm.LoadDialogDataAsync();
         }
     }
 }